requirements.txt
boto3==1.39.3
cryptography==44.0.0
fusepy==3.0.1